Almost finished with the animations.  I successfully got all of the data.  Not I just have to make them work.

+++ trunk/DirectX11 Engine/DirectX11 Engine/Src/Sound/SoundManager.cpp
@@ -360,7 +360,7 @@ bool SoundManager::PlayWaveFile()
 		return false;
 	}
 	// Set the volume
-	result = m_secondaryBuffer->SetVolume(m_fVolume);
+	result = m_secondaryBuffer->SetVolume((long)m_fVolume);
 	if(FAILED(result))
 	{
 		return false;
